Cambios a las vistas de dashboard y creado el Test

+++ src/main/java/acme/dashboard/Dashboard.java
@@ -19,8 +19,8 @@ public class Dashboard implements Serializable {
 	
 	Double		averageWorkload;
 	Double		standardDeviationWorkload;
-	Double		maxWorkload;
-	Double		minWorkload;
+	Integer		maxWorkload;
+	Integer		minWorkload;
 	
 	Double		averageExecutionPeriod;
 	Double		standardDeviationExecutionPeriod;
